Fix bug where coyote_timer and jump_buffer_timer would not work properly when set to 0 seconds
This commit adds the is_coyote_time_enabled and is_jump_buffer_enabled variables which set themselves to false if either of these are set to zero seconds. This avoids godot raising an error when setting wait_time to 0.
Now to start the coyote_timer, start_coyote_timer() must be called and is_coyote_timer_running() must be called instead of not coyote_timer.is_stopped(). This is the same for jump_buffer.